requirements.txt
python-docx
pandas
opencv-python
numpy
docx2txt